rag-based-chatbot/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.14",
    "huggingface-hub>=0.33.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "langchain-huggingface>=0.3.0",
    "openai>=1.92.2",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
